Finished the GUI, logic is all that needs to be implemented. Key listeners may be implemented in the future as well.
Added a bunch of methods in the CalculatorController to call methods in the DisplayPanel. Added the CalculatorMath class, all it does is return "42", will put logic in there. Added MouseListeners and turned ans into a CalculatorButton. Added handling for ans, and added a MouseListener. These MouseListeners detect clicks to do what the button should do, make the button look like it gets clicked, and makes the button change when the mouse goes over it. Added a getter to the CalculatorFrame for the CalculatorPanel. Added a getter to the CalculatorPanel for the DisplayPanel. Added handling for every button to the DisplayPanel.